EmphStrongTrait: reject fewer chars, better empty check
If rejecting a text block for strong, just reject the first marker so it can be tested again for emphasis. Explicitly check for blank chars at start and end of substring, because ctype_graph() is returning false for the first 8 bits of an extended UTF-8 char (this was causing some test cases to accidentally pass).